Move ProductsResult into a dedicated module
