+++ refurbished/feedback.py
@@ -1,13 +1,10 @@
 import csv
 import io
 import json
-from dataclasses import asdict, is_dataclass
-from typing import List
+from dataclasses import asdict
 
 import click
 from pydantic.json import pydantic_encoder
-
-from .model import Product
 
 
 class Feedback:
@@ -71,31 +68,3 @@ def echo(value, nl=True, err=False):
 
 def result(values):
     _current_feedback.result(values)
-
-
-class ProductsResult:
-    def __init__(self, values: List[Product]):
-        self.values = values
-
-    def str(self) -> str:
-        if len(self.values) == 0:
-            return "No products found"
-        out = ""
-        for p in self.values:
-            out += (
-                f"{p.previous_price} "
-                f"{p.price} "
-                f"{p.savings_price} "
-                f"({p.saving_percentage * 100}%) {p.name}\n"
-            )
-        return out
-
-    def data(self) -> List[Product]:
-        return self.values
-
-    def fieldnames(self) -> List[str]:
-        if self.values:
-            return (
-                self.values[0].__pydantic_model__.schema()["properties"].keys()
-            )
-        return []
